make Encoder/Decoder generate cpp classes
